How do I send a tweet to my server?
like this: I created a form that sends a tweet to my server. Todo: have my server forward the tweet to twitter somehow.

Analogy: A form is like a form letter.  You fill out the form, stick it in an envelope, and send it in the post.

pressing the submit button is like sticking a letter in the mailbox.  The mailman will come, pick up the letter, and deliver it.

A post request is like a box being sent through the mail. Post requests tend to carry a lot of data in the same way a box from your parents tends to carry a lot of cookies.

An action is like an address. An address tells a mailman where to deliver a letter. An action tells the internet where to deliver a post request.

In short, my form sends a tweet to my server in the same way I can send a box through the mail.