How do I show that I have sent a message to twitter?
Use an!  An ajax request is like sending a package in the mail and still being able to do work because you don't have to wait for the package to come back to work on something else.

A callback function is like a secretary giving you a message telling you that you got a reply to the letter you sent three weeks ago.