How do I add the ability to tweet to my sinatra application?
Analogy: the twitter gem is like a valet.  He takes your keys and takes your car and takes care of it for you while you go off to lunch.  This twitter app is the same way. you give it your tweet and you give the gem your keys, and it tweets for you.